Simplify markdown features
