+++ package.json
@@ -25,6 +25,8 @@
     "react": "17.0.2",
     "react-bootstrap": "^2.0.0-beta.0",
     "react-dom": "17.0.2",
+    "rehype-add-classes": "^1.0.0",
+    "rehype-autolink-headings": "^5.1.0",
     "rehype-slug": "^4.0.1",
     "remark-toc": "^7.2.0",
     "url-loader": "^4.1.1"
