+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import { AppProps } from 'next/app'
 import { useEffect } from 'react'
 import { useRouter } from 'next/router'
-import { Dropdown, Nav, Navbar, NavDropdown, Container } from  'react-bootstrap'
+import { Dropdown, Nav, Navbar, NavDropdown, Container } from 'react-bootstrap'
 // TODO: update these CSS imports once we settle on a CSS strategy.
 
 import 'bootstrap/dist/css/bootstrap.min.css'
