+++ pages/index.tsx
@@ -24,7 +24,7 @@ export default function Home(props: Props) {
     return (
         <PageLayout>
             <div className="row">
-                <img src="/headers/sourcegraph-learn-header.svg" className="w-100 mb-5"/>
+                <img src="/headers/sourcegraph-learn-header.svg" className="w-100 mb-5" />
             </div>
             <div className="row row-cols-2">
                 {props.posts.map(post => (
