+++ types.d.ts
@@ -1,1 +1,2 @@
 declare module '@mapbox/rehype-prism'
+declare module 'rehype-add-classes'
